"Feat: 꿀조합 CRUD 로직 일부 수정"

+++ src/main/java/com/zerobase/foodlier/module/recipe/domain/dto/RecipeDtoResponse.java
@@ -29,7 +29,7 @@ public class RecipeDtoResponse {
 
     private int expectedTime;
 
-    private Long heartCount;
+    private int heartCount;
 
     private Difficulty difficulty;
